refactor: use new EventHandler
- Introduce a new `Publisher` behavior with HTTP and UDP implemenations.

- Use the publisher in the new EventHandler

- Start new EventHandler and publisher-specific GenServers at start time.

- Remove now-obsolete protocol-specific EventHandlers.
